requirements.txt
telethon
psutil
pillow
wmi